move traits to separate file
