+++ src/lib.rs
@@ -36,30 +36,8 @@ use core::iter;
 
 mod error;
 pub use error::FromHexError;
-
-/// Encoding values as hex string.
-///
-/// This trait is implemented for all `T` which implement `AsRef<[u8]>`. This
-/// includes `String`, `str`, `Vec<u8>` and `[u8]`.
-///
-/// # Example
-///
-/// ```
-/// use hex::ToHex;
-///
-/// println!("{}", "Hello world!".encode_hex::<String>());
-/// ```
-///
-/// *Note*: instead of using this trait, you might want to use `encode()`.
-pub trait ToHex {
-    /// Encode the hex strict representing `self` into the result.. Lower case
-    /// letters are used (e.g. `f9b4ca`)
-    fn encode_hex<T: iter::FromIterator<char>>(&self) -> T;
-
-    /// Encode the hex strict representing `self` into the result.. Lower case
-    /// letters are used (e.g. `F9B4CA`)
-    fn encode_hex_upper<T: iter::FromIterator<char>>(&self) -> T;
-}
+mod traits;
+pub use traits::{FromHex, ToHex};
 
 const HEX_CHARS_LOWER: &[u8; 16] = b"0123456789abcdef";
 const HEX_CHARS_UPPER: &[u8; 16] = b"0123456789ABCDEF";
@@ -114,47 +92,6 @@ fn encode_to_iter<T: iter::FromIterator<char>>(table: &'static [u8; 16], source:
     BytesToHexChars::new(source, table).collect()
 }
 
-impl<T: AsRef<[u8]>> ToHex for T {
-    fn encode_hex<U: iter::FromIterator<char>>(&self) -> U {
-        encode_to_iter(HEX_CHARS_LOWER, self.as_ref())
-    }
-
-    fn encode_hex_upper<U: iter::FromIterator<char>>(&self) -> U {
-        encode_to_iter(HEX_CHARS_UPPER, self.as_ref())
-    }
-}
-
-/// Types that can be decoded from a hex string.
-///
-/// This trait is implemented for `Vec<u8>` and small `u8`-arrays.
-///
-/// # Example
-///
-/// ```
-/// use hex::FromHex;
-///
-/// match Vec::from_hex("48656c6c6f20776f726c6421") {
-///     Ok(vec) => {
-///         for b in vec {
-///             println!("{}", b as char);
-///         }
-///     }
-///     Err(e) => {
-///         // Deal with the error ...
-///     }
-/// }
-/// ```
-pub trait FromHex: Sized {
-    type Error;
-
-    /// Creates an instance of type `Self` from the given hex string, or fails
-    /// with a custom error type.
-    ///
-    /// Both, upper and lower case characters are valid and can even be
-    /// mixed (e.g. `f9b4ca`, `F9B4CA` and `f9B4Ca` are all valid strings).
-    fn from_hex<T: AsRef<[u8]>>(hex: T) -> Result<Self, Self::Error>;
-}
-
 fn val(c: u8, idx: usize) -> Result<u8, FromHexError> {
     match c {
         b'A'..=b'F' => Ok(c - b'A' + 10),
@@ -165,62 +102,6 @@ fn val(c: u8, idx: usize) -> Result<u8, FromHexError> {
             index: idx,
         }),
     }
-}
-
-impl FromHex for Vec<u8> {
-    type Error = FromHexError;
-
-    fn from_hex<T: AsRef<[u8]>>(hex: T) -> Result<Self, Self::Error> {
-        let hex = hex.as_ref();
-        if hex.len() % 2 != 0 {
-            return Err(FromHexError::OddLength);
-        }
-
-        hex.chunks(2)
-            .enumerate()
-            .map(|(i, pair)| Ok(val(pair[0], 2 * i)? << 4 | val(pair[1], 2 * i + 1)?))
-            .collect()
-    }
-}
-
-// Helper macro to implement the trait for a few fixed sized arrays. Once Rust
-// has type level integers, this should be removed.
-macro_rules! from_hex_array_impl {
-    ($($len:expr)+) => {$(
-        impl FromHex for [u8; $len] {
-            type Error = FromHexError;
-
-            fn from_hex<T: AsRef<[u8]>>(hex: T) -> Result<Self, Self::Error> {
-                let mut out = [0u8; $len];
-                decode_to_slice(hex, &mut out as &mut [u8])?;
-                Ok(out)
-            }
-        }
-    )+}
-}
-
-from_hex_array_impl! {
-    1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16
-    17 18 19 20 21 22 23 24 25 26 27 28 29 30 31 32
-    33 34 35 36 37 38 39 40 41 42 43 44 45 46 47 48
-    49 50 51 52 53 54 55 56 57 58 59 60 61 62 63 64
-    65 66 67 68 69 70 71 72 73 74 75 76 77 78 79 80
-    81 82 83 84 85 86 87 88 89 90 91 92 93 94 95 96
-    97 98 99 100 101 102 103 104 105 106 107 108 109 110 111 112
-    113 114 115 116 117 118 119 120 121 122 123 124 125 126 127 128
-    160 192 200 224 256 384 512 768 1024 2048 4096 8192 16384 32768
-}
-
-#[cfg(any(target_pointer_width = "32", target_pointer_width = "64"))]
-from_hex_array_impl! {
-    65536 131072 262144 524288 1048576 2097152 4194304 8388608
-    16777216 33554432 67108864 134217728 268435456 536870912
-    1073741824 2147483648
-}
-
-#[cfg(target_pointer_width = "64")]
-from_hex_array_impl! {
-    4294967296
 }
 
 /// Encodes `data` as hex string using lowercase characters.
